Changed module communication to use linking
The current solution for module to module communication causes
issues when communicating from any network using the same router
as the service itself.

This fix provides the builder (flaming-computing machine) with a
link to sitewatcher, named xyz to hide the internals.
The builder is then responsible for adding the link to every module
module built.